Filtering by year
Query Strings

Query strings can be included in the request URI in the following format: baseURL/path?query=value With query being the name of the query string, and value being its value.

For example, a request made to an API that manages a database of cars, to return only yellow cars, may be: findcars.com/cars?color=yellow The query string here is color and its value is yellow.
Filter by year

In Express, we can access these query strings in a route callback function using req.query.<queryString>, where <queryString> is the name of the value you are expecting.