Modifying GET/quotes route
adding the appropriate SQL statements to our GET '/quotes' route. For this route, we'll need to write a SELECT statement that returns all of the quotes from the database.

This route also accepts a query string to filter the results by year. If we receive a year to filter by, we'll use a SELECT statement with a WHERE clause to select quotes that match the given year.